server-py/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "pyautogui>=0.9.54",
    "pystray>=0.19.5",
    "Pillow>=10.0.0",
    "fastapi>=0.128.0",
    "uvicorn[standard]>=0.40.0",
    "zeroconf>=0.148.0",
    "ifaddr>=0.2.0",
    "loguru>=0.7.3",
    "pyperclip>=1.8.2",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pyinstaller>=6.17.0",
    "pytest>=9.0.2",
    "ruff>=0.14.10",
    "ty>=0.0.8",
]

[build-system]
requires = ["uv_build>=0.8.5,<0.10.0"]
build-backend = "uv_build"

[tool.ruff]
line-length = 100

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint]
select = [
    "E",   # pycodestyle
    "F",   # Pyflakes
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
    "I",   # isort
]

[tool.ty.environment]
python-version = "3.13"

[tool.ty.src]
include = [
    "src/**/*.py",
    "tests/**/*.py",
]
